Add aviation module
